refactor(lambda-tiler): relocated licensor attribution functions again
